requirements.txt
sanic==22.3.0
sanic-openapi==21.12.0
webargs-sanic==2.2.0
marshmallow==3.14.1
umongo==3.1.0
motor==2.5.1
sanic-jinja2==2022.1.13
sanic_session==0.8.0
aiohttp==3.8.1
aioredis==2.0.1
ujson==5.1.0
plotly==5.8.0
web3
